blub

+++ 01_download_data.py
@@ -16,6 +16,7 @@ output_dir.mkdir(parents=True, exist_ok=True)
 # Downloaded from: https://download.bls.gov/pub/time.series/cx/cx.series
 series_df = pd.read_csv(output_dir / "cx.series", sep="\t")
 series_df.columns = series_df.columns.str.strip()
+series_df["series_id"] = series_df["series_id"].str.strip()
 
 # Load characteristic codes (income quintiles)
 # Downloaded from: https://download.bls.gov/pub/time.series/cx/cx.characteristics
@@ -45,19 +46,22 @@ filtered_series = series_df[
 year_ranges = [(start, min(start + 9, 2025)) for start in range(1984, 2026, 10)]
 
 # Download and save each response as JSON
-for sid in filtered_series['series_id'].unique():
+for i_series_id in filtered_series['series_id'].unique():
     for start, end in year_ranges:
-        url = f"https://api.bls.gov/publicAPI/v2/timeseries/data/{sid}"
+        out_file = output_dir / "requests" / f"{i_series_id}_{start}_{end}.json"
+        if out_file.exists():
+            print(f"⏭️ Skipping (already exists): {out_file.name}")
+            continue
+
+        url = f"https://api.bls.gov/publicAPI/v2/timeseries/data/{i_series_id}"
         params = {"startyear": start, "endyear": end}
         r = requests.get(url, params=params)
-
-        out_file = output_dir / f"{sid}_{start}_{end}.json"
 
         if r.status_code == 200:
             with open(out_file, "w", encoding="utf-8") as f:
                 json.dump(r.json(), f, ensure_ascii=False, indent=2)
             print(f"✅ Saved JSON: {out_file.name}")
         else:
-            print(f"❌ Failed: {sid} ({start}-{end}) — HTTP {r.status_code}")
+            print(f"❌ Failed: {i_series_id} ({start}-{end}) — HTTP {r.status_code}")
         
-        time.sleep(0.5)
+        # time.sleep(0.)
